+++ src/components/feed/FeedOutputBrowser/FeedOutputBrowser.scss
@@ -146,6 +146,10 @@
       height: 100%;
       overflow: scroll;
     }
+
+    .object-key {
+      color: white !important;
+    }
   }
 
   .dcm-preview {
@@ -177,14 +181,5 @@
 }
 
 .overlay {
-  /* prevent text selection on overlay
-  -webkit-touch-callout: none;
-  -webkit-user-select: none;
-  -khtml-user-select: none;
-  -moz-user-select: none;
-  -ms-user-select: none;
-  user-select: none;
-
-  /* ignore pointer event on overlay */
   pointer-events: none;
 }
